remove d3 topology
