--- frontend/src/components/Topology/css/diagram-controls.css
@@ -1,40 +1,0 @@
-.diagram-controls {
-    position: absolute;
-    left: 0px;
-    top: 0px;
-    display: flex;
-    justify-content: center;
-    text-align: center;
-}
-.diagram-controls .rectangle-zoom {
-    position: relative;
-    background-color: var(--pf-global--Color--light-300);
-    border-radius: 3px;
-    box-shadow: var(--pf-c-page--section--m-sticky-top--BoxShadow);
-    width: 30px;
-    height: 30px;
-    margin-right: 5px;
-    display: flex;
-    text-align: center;
-    justify-content: center;
-}
-.diagram-controls .zoom-in-button {
-    position: relative;
-}
-.diagram-controls .zoom-out-button {
-    position: relative;
-    text-align: center;
-}
-.diagram-controls .zoom-target-button {
-    position: relative;
-}
-.diagram-controls .icon {
-    cursor: pointer;
-    width: 18px;
-    height: 22px;
-    display: block;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
