--- frontend/src/components/Topology/css/diagram-filter-bar.css
@@ -1,31 +1,0 @@
-.filter-bar {
-    display: flex;
-    flex-direction: row;
-    padding: 4px 0 2px;
-}
-.filter-bar .filter-bar-button-container {
-    margin: 0 4px;
-    position: relative;
-    cursor: pointer;
-}
-.filter-bar .filter-bar-button-container .filter-bar-button {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding: 2px;
-}
-.filter-bar .filter-bar-button-container .filter-bar-button .filter-bar-button-icon {
-    stroke-width: 3px;
-    fill: #fff;
-}
-.filter-bar .filter-bar-button-container .filter-bar-button .filter-bar-button-label {
-    font-size: 0.8rem;
-    margin-left: 3px;
-}
-.filter-bar .filter-bar-button-container .filter-bar-button-checkmark {
-    pointer-events: none;
-    position: absolute;
-    top: -4px;
-    left: 12px;
-    fill: blue;
-}
