--- frontend/src/components/Topology/css/resource-toolbar.css
@@ -1,65 +1,0 @@
-.resource-toolbar {
-    top: 3.082rem;
-    margin-bottom: 8px;
-    flex-direction: column;
-    align-items: flex-end;
-}
-.resource-toolbar.hide {
-    display: none;
-}
-.resource-toolbar .resource-toolbar-container {
-    position: relative;
-    top: 0px;
-    left: 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-.resource-toolbar .resource-toolbar-container > div {
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-}
-.resource-toolbar .resource-toolbar-container .resource-toolbar-buttons {
-    height: 0;
-    margin: 0px 0px 3px 0px;
-}
-.resource-toolbar .resource-toolbar-container .resource-toolbar-buttons > div {
-    color: #06c;
-    font-size: 14px;
-    font-weight: bold;
-    cursor: pointer;
-    outline: none;
-    display: flex;
-    align-items: center;
-    height: 26px;
-}
-.resource-toolbar .resource-toolbar-container .resource-toolbar-buttons > div .button-icon {
-    width: 12px;
-    height: 20px;
-    fill: #06c;
-    stroke: none;
-}
-.resource-toolbar .resource-toolbar-container .resource-toolbar-buttons > div .button-label {
-    padding: 0 8px;
-    white-space: nowrap;
-}
-.resource-toolbar .resource-toolbar-container .resource-toolbar-buttons > div:not(:last-child) {
-    padding-right: 20px;
-    border-right: 2px solid #dfe3e6;
-}
-.resource-toolbar .resource-toolbar-container .resource-toolbar-buttons > div:last-child {
-    padding-left: 20px;
-}
-.resource-toolbar .resource-toolbar-container .resource-filter-bar .pf-c-chip-group.pf-m-category {
-    background-color: #fff;
-}
-.resource-toolbar .resource-toolbar-container .resource-filter-bar .pf-c-chip-group__label {
-    padding: 0 10px;
-}
-.resource-toolbar .resource-toolbar-container .resource-filter-bar .pf-c-chip__text {
-    overflow: visible;
-}
-.resource-toolbar .resource-toolbar-container .resource-filter-bar .pf-c-chip-group__close {
-    padding: 3px 0;
-}
