--- frontend/src/components/Topology/css/topology-link.css
@@ -1,21 +1,0 @@
-.link path {
-    fill: none;
-    stroke: #8c9ba5;
-    stroke-width: 1px;
-}
-.arrowDecorator {
-    stroke: none;
-    fill: #42535c;
-}
-.arrowDecorator.faded {
-    opacity: 0.3;
-}
-.squareDecorator {
-    stroke: none;
-    fill: #8c9ba5;
-}
-.label text {
-    text-anchor: middle;
-    font-size: 0.6rem;
-    fill: #394b54;
-}
