--- frontend/src/components/Topology/shapes/DiagramShapes.tsx
@@ -1,770 +1,0 @@
-/* Copyright Contributors to the Open Cluster Management project */
-export const DiagramShapes = () => (
-    <svg xmlns="http://www.w3.org/2000/svg" xmlnsXlink="http://www.w3.org/1999/xlink" width="0" height="0">
-        <symbol viewBox="0 0 36 36" id="diagramShapes_select">
-            <circle cx="18" cy="18" r="16" strokeWidth="2" />
-        </symbol>
-        <symbol viewBox="0 0 64 46" id="diagramIcons_selectMultiplier">
-            <rect x="2" y="2" width="60px" height="44px" rx="22px" ry="22px" />
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_multiplier">
-            <circle cx="25.795" cy="15.795" r="15.093" fill="#F0F0F0" stroke="#c7c7c7" strokeWidth="1" />
-            <circle cx="20.795" cy="15.795" r="15.093" fill="#F0F0F0" stroke="#c7c7c7" strokeWidth="1" />
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_application">
-            <g id="Application" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g id="02_Software/App/App_outline" transform="matrix(1.6379 0 0 1.6379 7.9512 7.8634)" stroke="none">
-                    <path
-                        id="path-application"
-                        d="m9.1579 0.38865c0.18504 0 0.33505 0.15 0.33505 0.33505v8.3806c0 0.18504-0.15 0.33505-0.33505 0.33505h-8.4878c-0.18504 0-0.33505-0.15-0.33505-0.33505v-8.3806c0-0.08886 0.035299-0.17408 0.098132-0.23691 0.062833-0.062833 0.14805-0.098132 0.23691-0.098132zm-0.33505 0.67009h-7.8177v7.7105h7.8177zm-1.1347 0.70583c0.1253-0.052687 0.27001-0.02445 0.36632 0.071476 0.030284 0.030984 0.054524 0.067344 0.071476 0.10721 0.017815 0.04086 0.026942 0.084977 0.026805 0.12955 0 0.18504-0.15001 0.33505-0.33505 0.33505-0.044574 1.372e-4 -0.088691-0.00899-0.12955-0.026805-0.03987-0.016953-0.07623-0.041192-0.10721-0.071476-0.062851-0.062773-0.098202-0.14794-0.098282-0.23677-1.373e-4 -0.044574 0.00899-0.088691 0.026805-0.12955 0.033548-0.080878 0.097813-0.14514 0.17869-0.17869zm-0.84154-0.019773c0.10926-0.021494 0.22206 0.012702 0.301 0.091249 0.094167 0.09407 0.12372 0.23496 0.075289 0.35894-0.04843 0.12398-0.16566 0.20754-0.29865 0.21287h-0.013402c-0.18504 0-0.33505-0.15-0.33505-0.33505-1.373e-4 -0.044574 0.00899-0.088691 0.026805-0.12955 0.043301-0.10259 0.13474-0.17697 0.24401-0.19846z"
-                        fill="#000"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_ansiblejob">
-            <g transform="translate(0)" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.4248 0 0 1.4248 6.7502 6.7876)" fillRule="nonzero" stroke="none">
-                    <ellipse
-                        id="Oval"
-                        transform="rotate(-80.819 6.492 6.4945)"
-                        cx="6.492"
-                        cy="6.4945"
-                        rx="5.6444"
-                        ry="5.6141"
-                        fill="#000"
-                    />
-                    <path
-                        id="Shape"
-                        d="m9.1463 8.7491-2.2284-5.392c-0.052056-0.1455-0.19075-0.24126-0.34446-0.23784-0.15659-0.0049358-0.29905 0.090577-0.35475 0.23784l-2.4458 5.915h0.83655l0.96802-2.4375 2.8886 2.3474c0.11678 0.094545 0.20051 0.13739 0.30921 0.13739 0.10774 0.00259 0.21195-0.0387 0.28898-0.1145 0.07703-0.075799 0.12035-0.17968 0.12012-0.28806-0.00333-0.057643-0.016244-0.11432-0.038195-0.16767zm-2.5706-4.4237 1.4462 3.5949-2.188-1.7328z"
-                        fill="#fff"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_chart">
-            <g id="HelmChart" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <path
-                    id="Combined-Shape"
-                    d="m21.265 11.489c0.9315 1.0862 1.5327 2.464 1.6478 3.9777h0.55354c0.29455 0 0.53333 0.23878 0.53333 0.53333 0 0.29455-0.23878 0.53333-0.53333 0.53333h-0.55354c-0.11515 1.5137-0.71633 2.8915-1.6478 3.9777l0.39156 0.39156c0.20828 0.20828 0.20828 0.54597 0 0.75425-0.20828 0.20828-0.54597 0.20828-0.75425 0l-0.39156-0.39156c-1.0862 0.9315-2.464 1.5327-3.9777 1.6478v0.55354c0 0.29455-0.23878 0.53333-0.53333 0.53333-0.29455 0-0.53333-0.23878-0.53333-0.53333v-0.55354c-1.5137-0.11515-2.8915-0.71633-3.9777-1.6478l-0.39156 0.39156c-0.20828 0.20828-0.54597 0.20828-0.75425 0-0.20828-0.20828-0.20828-0.54597 0-0.75425l0.39156-0.39156c-0.9315-1.0862-1.5327-2.464-1.6478-3.9777h-0.55354c-0.29455 0-0.53333-0.23878-0.53333-0.53333 0-0.29455 0.23878-0.53333 0.53333-0.53333h0.55354c0.11515-1.5137 0.71633-2.8915 1.6478-3.9777l-0.39156-0.39156c-0.20828-0.20828-0.20828-0.54597 0-0.75425 0.20828-0.20828 0.54597-0.20828 0.75425 0l0.39156 0.39156c1.0862-0.9315 2.464-1.5327 3.9777-1.6478v-0.55354c0-0.29455 0.23878-0.53333 0.53333-0.53333 0.29455 0 0.53333 0.23878 0.53333 0.53333v0.55354c1.5137 0.11515 2.8915 0.71633 3.9777 1.6478l0.39156-0.39156c0.20828-0.20828 0.54597-0.20828 0.75425 0 0.20828 0.20828 0.20828 0.54597 0 0.75425zm-9.7737 8.2652 3.2208-3.2208h-4.5552c0.10991 1.2196 0.59305 2.3316 1.3343 3.2208zm0.75425 0.75425c0.88925 0.74129 2.0012 1.2244 3.2208 1.3343v-4.5552zm8.2626-8.2626-3.2208 3.2208h4.5552c-0.10991-1.2196-0.59305-2.3316-1.3343-3.2208zm-0.75425-0.75425c-0.88925-0.74129-2.0012-1.2244-3.2208-1.3343v4.5552zm-8.2626 0.75425c-0.74129 0.88925-1.2244 2.0012-1.3343 3.2208h4.5552zm0.75425-0.75425 3.2208 3.2208v-4.5552c-1.2196 0.10991-2.3316 0.59305-3.2208 1.3343zm8.2626 8.2626c0.74129-0.88925 1.2244-2.0012 1.3343-3.2208h-4.5552zm-0.75425 0.75425-3.2208-3.2208v4.5552c1.2196-0.10991 2.3316-0.59305 3.2208-1.3343zm-3.7542-3.4417c0.5891 0 1.0667-0.47756 1.0667-1.0667s-0.47756-1.0667-1.0667-1.0667-1.0667 0.47756-1.0667 1.0667 0.47756 1.0667 1.0667 1.0667zm0-0.53333c-0.29455 0-0.53333-0.23878-0.53333-0.53333 0-0.29455 0.23878-0.53333 0.53333-0.53333 0.29455 0 0.53333 0.23878 0.53333 0.53333 0 0.29455-0.23878 0.53333-0.53333 0.53333z"
-                    fill="#171717"
-                    fillRule="nonzero"
-                    stroke="none"
-                />
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_cluster">
-            <g id="Cluster" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.8606 0 0 1.8606 6.8571 6.8572)" fill="#000000" stroke="none">
-                    <path
-                        id="Fill"
-                        d="m7.9852 6.7568c-0.2187 0.0011364-0.43304 0.0613-0.62039 0.17414l-1.1711-1.1714c0.34117-0.51205 0.34117-1.179 0-1.691l0.78225-0.78225c0.40859 0.19681 0.89955 0.070194 1.162-0.29966 0.26245-0.36986 0.21985-0.87509-0.10084-1.1958-0.32068-0.32068-0.82591-0.36329-1.1958-0.10084-0.36986 0.26245-0.49647 0.75341-0.29966 1.162l-0.78225 0.78225c-0.51205-0.34117-1.179-0.34117-1.691 0l-1.1714-1.1714c0.11279-0.18726 0.17295-0.40149 0.17414-0.62009 0-0.67848-0.55002-1.2285-1.2285-1.2285s-1.2285 0.55002-1.2285 1.2285 0.55002 1.2285 1.2285 1.2285c0.2187-0.00114 0.43304-0.0613 0.62039-0.17414l1.1711 1.1714c-0.34117 0.51205-0.34117 1.179 0 1.691l-1.1714 1.1714c-0.18727-0.11277-0.40149-0.17293-0.62009-0.17414-0.67848 0-1.2285 0.55002-1.2285 1.2285s0.55002 1.2285 1.2285 1.2285 1.2285-0.55002 1.2285-1.2285c-0.0011364-0.2187-0.0613-0.43304-0.17414-0.62039l1.1714-1.1711c0.16334 0.10887 0.34614 0.1852 0.53839 0.22482v1.0089c-0.42308 0.14958-0.67686 0.58221-0.60098 1.0245 0.075883 0.44228 0.45936 0.76557 0.90811 0.76557s0.83222-0.32329 0.90811-0.76557c0.075883-0.44228-0.1779-0.87491-0.60098-1.0245v-1.0089c0.19225-0.039617 0.37505-0.11595 0.53839-0.22482l1.1714 1.1714c-0.11279 0.18726-0.17295 0.40149-0.17414 0.62009 0 0.67848 0.55002 1.2285 1.2285 1.2285s1.2285-0.55002 1.2285-1.2285-0.55002-1.2285-1.2285-1.2285zm-3.0712-2.7641c0.50886 0 0.92138 0.41251 0.92138 0.92138s-0.41251 0.92138-0.92138 0.92138-0.92138-0.41251-0.92138-0.92138c5.079e-4 -0.50865 0.41272-0.92087 0.92138-0.92138zm-3.6855-2.1499c0-0.33924 0.27501-0.61425 0.61425-0.61425s0.61425 0.27501 0.61425 0.61425-0.27501 0.61425-0.61425 0.61425c-0.3391-3.386e-4 -0.61391-0.27515-0.61425-0.61425zm0.61425 6.7568c-0.33924 0-0.61425-0.27501-0.61425-0.61425s0.27501-0.61425 0.61425-0.61425 0.61425 0.27501 0.61425 0.61425c-3.386e-4 0.3391-0.27515 0.61391-0.61425 0.61425zm6.1425 0c-0.33924 0-0.61425-0.27501-0.61425-0.61425s0.27501-0.61425 0.61425-0.61425 0.61425 0.27501 0.61425 0.61425c-5.074e-4 0.33903-0.27522 0.61374-0.61425 0.61425z"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_configmap">
-            <g id="ConfigMap" fill="none" fillRule="evenodd" stroke="none">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g
-                    id="Organization-/-File-/-catalog-/-32"
-                    transform="matrix(2.035002,0,0,2.035002,5.6582201,5.9193449)"
-                    fill="#000000"
-                >
-                    <path
-                        d="M 7.98525,0.61425 H 2.457 c -0.3392409,0 -0.61425,0.27500909 -0.61425,0.61425 V 2.457 H 1.2285 V 3.07125 H 1.84275 V 4.606875 H 1.2285 v 0.61425 H 1.84275 V 6.75675 H 1.2285 V 7.371 h 0.61425 v 1.2285 c 0,0.3392409 0.2750091,0.61425 0.61425,0.61425 h 5.52825 c 0.3392409,0 0.61425,-0.2750091 0.61425,-0.61425 v -7.371 c 0,-0.33924091 -0.2750091,-0.61425 -0.61425,-0.61425 z m 0,7.98525 H 2.457 V 7.371 H 3.07125 V 6.75675 H 2.457 V 5.221125 H 3.07125 V 4.606875 H 2.457 V 3.07125 H 3.07125 V 2.457 H 2.457 V 1.2285 H 7.98525 Z M 4.29975,2.457 h 2.457 v 0.61425 h -2.457 z m 0,2.149875 h 2.457 v 0.61425 h -2.457 z m 0,2.149875 h 2.457 V 7.371 h -2.457 z"
-                        id="Fill"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_container">
-            <g id="Container" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g
-                    id="02_Software/Containerized-app/Containerized-app_outline"
-                    transform="matrix(1.6604 0 0 1.6604 7.8713 6.5104)"
-                    stroke="none"
-                >
-                    <path
-                        id="path-container"
-                        d="m4.7222 0.33467c0.095139-0.053051 0.21097-0.053051 0.3061 0l4.514 2.6039c0.10126 0.050539 0.16695 0.15221 0.17142 0.26529v5.2119c-3.3e-4 0.10879-0.057634 0.20946-0.15101 0.26529l-4.514 2.6039c-0.045644 0.02719-0.097885 0.04131-0.15101 0.04083-0.054406 4.07e-4 -0.10794-0.01368-0.15509-0.04083l-4.5099-2.6039c-0.096184-0.053555-0.15561-0.1552-0.1551-0.26529v-5.2119c-5.1234e-4 -0.11009 0.058913-0.21174 0.1551-0.26529zm0.1755 0.62037-4.2283 2.4284v4.8528l4.2283 2.4284 4.1834-2.4203 0.020407-4.8609zm2.0407 2.6039c0.16814 0.0022 0.30391 0.13797 0.3061 0.3061v3.9141c0 0.16906-0.13705 0.3061-0.3061 0.3061h-3.963c-0.16906 0-0.3061-0.13705-0.3061-0.3061v-3.9141c0.0022-0.16814 0.13797-0.30391 0.3061-0.3061zm-0.3061 0.61221h-3.3508l-0.016326 3.3018h3.3672zm-1.8325 0.57956c0.11951-0.11936 0.31312-0.11936 0.43263 0 0.027793 0.029756 0.049923 0.064334 0.065302 0.10203 0.01584 0.036034 0.024172 0.074918 0.024488 0.11428 8.534e-4 0.081339-0.031587 0.15949-0.089791 0.21631-0.028143 0.028912-0.061342 0.052427-0.097953 0.069384-0.038231 0.012602-0.078117 0.019479-0.11836 0.020407-0.081157-7.15e-5 -0.15896-0.032368-0.21631-0.089791-0.028621-0.027523-0.050905-0.060948-0.065302-0.097954-0.016276-0.03733-0.024615-0.077636-0.02449-0.11836 3.18e-4 -0.039361 0.00865-0.078245 0.02449-0.11428 0.013865-0.038466 0.03618-0.073333 0.065302-0.10203zm0.81628 0c0.12206-0.11304 0.31057-0.11304 0.43263 0 0.027793 0.029756 0.049923 0.064334 0.065302 0.10203 0.01584 0.036034 0.024172 0.074918 0.02449 0.11428 1.254e-4 0.040724-0.00821 0.08103-0.02449 0.11836-0.015488 0.036427-0.037634 0.069646-0.065302 0.097954-0.11951 0.11936-0.31312 0.11936-0.43263 0-0.02905-0.027162-0.051403-0.060692-0.065302-0.097954-0.016276-0.03733-0.024615-0.077636-0.02449-0.11836 3.179e-4 -0.039361 0.00865-0.078245 0.02449-0.11428 0.013865-0.038466 0.03618-0.073333 0.065302-0.10203z"
-                        fill="#000"
-                        fillRule="evenodd"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_customresource">
-            <g id="CustomResource" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.6446 0 0 1.6446 7.7571 7.9767)" stroke="none" fill="#000000">
-                    <g id="Actions-/-Controls-/-settings-/-32-Copy">
-                        <path
-                            id="Fill"
-                            d="m8.9471 5.2823v-0.54756l0.68713-0.60124c0.26151-0.2305 0.31894-0.61587 0.136-0.9126l-0.8446-1.4315c-0.12773-0.22123-0.36368-0.35762-0.61914-0.35788-0.077786-5.959e-4 -0.15515 0.011492-0.22904 0.035788l-0.86965 0.29346c-0.15015-0.099775-0.30677-0.18944-0.46883-0.26841l-0.18252-0.90186c-0.067546-0.34007-0.36911-0.58283-0.71576-0.57619h-1.6749c-0.34665-0.0066431-0.64822 0.23612-0.71576 0.57619l-0.18252 0.90186c-0.16322 0.078949-0.32103 0.16862-0.4724 0.26841l-0.85176-0.30778c-0.074691-0.01946-0.15204-0.026712-0.22904-0.021473-0.25546 2.62e-4 -0.49141 0.13665-0.61914 0.35788l-0.8446 1.4315c-0.17228 0.29597-0.11078 0.67252 0.14673 0.89828l0.6764 0.61914v0.54756l-0.6764 0.60124c-0.2651 0.22761-0.32712 0.61336-0.14673 0.9126l0.8446 1.4315c0.12773 0.22123 0.36368 0.35762 0.61914 0.35788 0.077786 5.959e-4 0.15515-0.011492 0.22904-0.035788l0.86965-0.29346c0.15015 0.099775 0.30677 0.18944 0.46883 0.26841l0.18252 0.90186c0.067546 0.34007 0.36911 0.58283 0.71576 0.57619h1.6892c0.34665 0.006643 0.64822-0.23612 0.71576-0.57619l0.18252-0.90186c0.16322-0.078949 0.32103-0.16862 0.4724-0.26841l0.86608 0.29346c0.073896 0.024296 0.15126 0.036384 0.22904 0.035788 0.25546-2.62e-4 0.49141-0.13665 0.61914-0.35788l0.81597-1.4315c0.17228-0.29597 0.11078-0.67252-0.14673-0.89828zm-0.64061 2.5911-1.2275-0.41514c-0.28736 0.2434-0.61577 0.43366-0.96986 0.56188l-0.2541 1.2848h-1.6892l-0.2541-1.2705c-0.35129-0.13185-0.67793-0.32179-0.96628-0.56188l-1.2311 0.40083-0.8446-1.4315 0.97344-0.85892c-0.066174-0.37046-0.066174-0.74971 0-1.1202l-0.97344-0.88397 0.8446-1.4315 1.2275 0.41514c0.28736-0.2434 0.61577-0.43366 0.96986-0.56188l0.2541-1.2848h1.6892l0.2541 1.2705c0.35129 0.13185 0.67793 0.32179 0.96628 0.56188l1.2311-0.40083 0.8446 1.4315-0.97344 0.85892c0.066174 0.37046 0.066174 0.74971 0 1.1202l0.97344 0.88397z"
-                        />
-                    </g>
-                    <g transform="translate(.71576 .71576)">
-                        <path d="m4.2946 1.0736c1.7789 0 3.2209 1.4421 3.2209 3.2209s-1.4421 3.2209-3.2209 3.2209-3.2209-1.4421-3.2209-3.2209 1.4421-3.2209 3.2209-3.2209zm0-0.53682c-2.0754 0-3.7578 1.6824-3.7578 3.7578 0 2.0754 1.6824 3.7578 3.7578 3.7578 2.0754 0 3.7578-1.6824 3.7578-3.7578 0-2.0754-1.6824-3.7578-3.7578-3.7578zm1.6105 3.4894h-1.3421v-1.3421h-0.53682v1.3421h-1.3421v0.53682h1.3421v1.3421h0.53682v-1.3421h1.3421z" />
-                    </g>
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_daemonset">
-            <g id="DaemonSet" transform="translate(0)" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.7806 0 0 1.7806 8.625 8.625)">
-                    <g fill="#000" fillRule="nonzero" stroke="none">
-                        <g id="icon/file/template/32-copy" transform="translate(1.053,1.053)">
-                            <polygon
-                                id="Polygon"
-                                points="0.09405 1.053 0.09405 0.351 0.702 0 1.3099 0.351 1.3099 1.053 0.702 1.404"
-                            />
-                        </g>
-                        <g transform="translate(1.053,5.967)">
-                            <polygon points="0.09405 1.053 0.09405 0.351 0.702 0 1.3099 0.351 1.3099 1.053 0.702 1.404" />
-                        </g>
-                        <rect x="3.159" width="1" height="3.159" />
-                        <rect x="3.159" y="4.914" width="1" height="3.159" />
-                    </g>
-                    <rect id="Rectangle" width="7.722" height="3.51" rx=".702" stroke="#000" strokeWidth=".702" />
-                    <rect
-                        id="Rectangle-Copy"
-                        y="4.914"
-                        width="7.722"
-                        height="3.51"
-                        rx=".702"
-                        stroke="#000"
-                        strokeWidth=".702"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_deployment">
-            <g id="Deployment" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g
-                    id="Shape"
-                    transform="matrix(1.7533 0 0 1.7533 10.619 7.4743)"
-                    fillRule="nonzero"
-                    stroke="none"
-                    fill="#000000"
-                >
-                    <path d="m3.7294 0-2.2376 2.2815 0.52584 0.53615 1.3388-1.3613v4.6276h0.74588v-4.6276l1.3388 1.3613 0.52584-0.53615zm-3.7294 5.3235h2.6106v0.7605h-2.6106zm0 1.521h5.967v0.7605h-5.967zm0 1.521h5.967v0.7605h-5.967z" />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_deploymentconfig">
-            <g id="DeploymentConfig" transform="translate(0)" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.3031 0 0 1.3031 -4.7627 -4.5173)">
-                    <circle
-                        id="path-dc"
-                        cx="15.97"
-                        cy="15.97"
-                        r="5.4405"
-                        stroke="#000"
-                        strokeDasharray="2.106, 2.808"
-                        strokeWidth="1.404"
-                    />
-                    <g fill="#000" fillRule="nonzero" stroke="none">
-                        <polygon
-                            id="Shape-path"
-                            points="14.742 14.284 15.041 14.589 15.802 13.815 15.802 16.447 16.226 16.447 16.226 13.815 16.988 14.589 17.287 14.284 16.014 12.987"
-                        />
-                        <polygon points="17.901 17.691 14.04 17.691 14.04 17.199 17.901 17.199" />
-                        <polygon points="17.901 18.675 14.04 18.675 14.04 18.183 17.901 18.183" />
-                    </g>
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_host">
-            <g id="Host" transform="translate(0)" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g
-                    id="01_Hardware/Server-stack/Server-stack_outline"
-                    transform="matrix(1.7941 0 0 1.7941 7.6339 7.0305)"
-                    stroke="none"
-                >
-                    <path
-                        id="path-host"
-                        d="m8.7876 6.916c0.17179 0 0.31105 0.14103 0.31105 0.315 0.00415 2.226 0.00415 2.226-0.09124 2.3226-0.058277 0.059091-0.13734 0.092326-0.21981 0.0924h-8.2946c-0.17179 0-0.31105-0.14103-0.31105-0.315v-2.1c0-0.17397 0.13926-0.315 0.31105-0.315zm-0.31105 0.63h-7.6725v1.47h7.6725zm-1.0256 0.48297c0.10107-0.042384 0.21772-0.019324 0.2951 0.058336 0.078044 0.078202 0.10137 0.19573 0.059124 0.29785-0.04225 0.10212-0.14177 0.16874-0.25221 0.16884-0.035181 1.48e-4 -0.069978-0.00731-0.10201-0.021878-0.034078-0.012904-0.065097-0.032771-0.091079-0.058336-0.10655-0.10676-0.10655-0.27972 0-0.38648 0.025982-0.025565 0.057-0.045433 0.091079-0.058337zm-4.8423-0.020971c0.16753 0 0.30333 0.12223 0.30333 0.273s-0.13581 0.273-0.30333 0.273h-1.2133c-0.16753 0-0.30333-0.12223-0.30333-0.273s0.13581-0.273 0.30333-0.273zm6.1802-4.368c0.17088 0.002262 0.30886 0.14198 0.31109 0.315v2.1c-0.002234 0.17302-0.14022 0.31274-0.31109 0.315h-8.2958c-0.17181 0-0.31109-0.14103-0.31109-0.315v-2.1c0-0.17397 0.13928-0.315 0.31109-0.315zm-0.31109 0.63h-7.6736v1.47h7.6736zm-1.1512 0.583c0.07264-0.10798 0.21134-0.14991 0.33171-0.10029 0.032587 0.013843 0.062304 0.033637 0.087627 0.058365 0.078214 0.078241 0.1016 0.19583 0.059253 0.298-0.042343 0.10217-0.14208 0.16883-0.25276 0.16893-0.035258 1.481e-4 -0.070131-0.0073181-0.10223-0.021889-0.034153-0.01291-0.065239-0.032788-0.091278-0.058365-0.091446-0.09262-0.10496-0.23677-0.032322-0.34475zm-4.7179-0.121c0.16753 0 0.30333 0.12223 0.30333 0.273s-0.13581 0.273-0.30333 0.273h-1.2133c-0.16753 0-0.30333-0.12223-0.30333-0.273s0.13581-0.273 0.30333-0.273zm6.1802-4.368c0.17088 0.002262 0.30886 0.14198 0.31109 0.315v2.1c-0.002234 0.17302-0.14022 0.31274-0.31109 0.315h-8.2958c-0.17181 0-0.31109-0.14103-0.31109-0.315v-2.1c0-0.17397 0.13928-0.315 0.31109-0.315zm-0.31109 0.63h-7.6736v1.47h7.6736zm-1.1195 0.53929c0.0801-0.079068 0.20199-0.099711 0.30421-0.051521 0.029845 0.01619 0.056423 0.037655 0.078409 0.063326 0.071513 0.065778 0.10095 0.16484 0.076764 0.25833-0.024187 0.093487-0.09814 0.16649-0.19285 0.19036s-0.19507-0.005183-0.2617-0.075775c-0.024161-0.024401-0.043499-0.053035-0.057024-0.084434-0.048819-0.1009-0.027906-0.22122 0.052194-0.30029zm-4.7496-0.07729c0.16753 0 0.30333 0.12223 0.30333 0.273s-0.13581 0.273-0.30333 0.273h-1.2133c-0.16753 0-0.30333-0.12223-0.30333-0.273s0.13581-0.273 0.30333-0.273z"
-                        fill="#000"
-                        fillRule="evenodd"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_ingress">
-            <g id="Ingress" transform="translate(0)" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.3371 0 0 1.3371 8.4693 10.969)" stroke="none">
-                    <path
-                        id="path-ingress"
-                        d="m7.488 5.7868v1.0205c0 0.37596-0.30477 0.68073-0.68073 0.68073h-6.1265c-0.37596 0-0.68073-0.30477-0.68073-0.68073v-6.1265c0-0.37596 0.30477-0.68073 0.68073-0.68073h6.1265c0.37596 0 0.68073 0.30477 0.68073 0.68073v1.0236"
-                        stroke="#000"
-                        strokeWidth=".702"
-                    />
-                    <polygon
-                        transform="matrix(0 1 1 0 5.3182 -5.3182)"
-                        points="8.7218 6.2967 9.4025 6.2967 9.4025 2.4949 10.624 3.7134 11.104 3.2335 9.0622 1.1913 7.02 3.2335 7.4999 3.7134 8.7218 2.4949"
-                        fill="#000"
-                        fillRule="nonzero"
-                    />
-                    <path
-                        id="Combined-Shape"
-                        d="m3.744 2.3825c0.75191 0 1.3615 0.60954 1.3615 1.3615 0 0.36108-0.14344 0.70737-0.39876 0.96269-0.25532 0.25532-0.60161 0.39876-0.96269 0.39876-0.75191 0-1.3615-0.60954-1.3615-1.3615s0.60954-1.3615 1.3615-1.3615zm0 0.68073c-0.37596 0-0.68073 0.30477-0.68073 0.68073s0.30477 0.68073 0.68073 0.68073 0.68073-0.30477 0.68073-0.68073-0.30477-0.68073-0.68073-0.68073z"
-                        fill="#000"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_namespace">
-            <g id="Topology" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-                <g id="Artboard" transform="translate(-1245.000000, -1091.000000)">
-                    <g id="Namespace" transform="translate(1245.000000, 1091.000000)">
-                        <circle
-                            id="Oval-Copy-38"
-                            stroke="#c7c7c7"
-                            strokeWidth="1.404"
-                            fill="#FFFFFF"
-                            cx="15.795"
-                            cy="15.795"
-                            r="15.093"
-                        ></circle>
-                        <circle id="Oval-Copy-37" fill="#FAFAFA" cx="15.795" cy="15.795" r="10.53"></circle>
-                        <g
-                            id="Organization-/-File-/-categories-/-32"
-                            transform="translate(10.881000, 10.881000)"
-                            fill="#000000"
-                        >
-                            <path
-                                d="M2.076165,1.84275 L2.21437125,2.11609125 L2.38329,2.457 L3.6855,2.457 L3.6855,3.992625 L1.2285,3.992625 L1.2285,1.84275 L2.076165,1.84275 L2.076165,1.84275 Z M2.2665825,1.22849252 L0.921375,1.22849252 C0.751754546,1.22849252 0.61425,1.36600455 0.61425,1.535625 L0.61425,4.29975 C0.61425,4.46937045 0.751754546,4.606875 0.921375,4.606875 L3.992625,4.606875 C4.16224545,4.606875 4.29975,4.46937045 4.29975,4.29975 L4.29975,2.149875 C4.29975,1.98025455 4.16224545,1.84275 3.992625,1.84275 L2.764125,1.84275 L2.542995,1.39741875 C2.4904565,1.29314865 2.38333818,1.22768745 2.2665825,1.22849252 Z M6.990165,1.84275 L7.12837125,2.11609125 L7.29729,2.457 L8.5995,2.457 L8.5995,3.992625 L6.1425,3.992625 L6.1425,1.84275 L6.990165,1.84275 L6.990165,1.84275 Z M7.1805825,1.22849252 L5.835375,1.22849252 C5.66575455,1.22849252 5.52825,1.36600455 5.52825,1.535625 L5.52825,4.29975 C5.52825,4.46937045 5.66575455,4.606875 5.835375,4.606875 L8.906625,4.606875 C9.07624545,4.606875 9.21375,4.46937045 9.21375,4.29975 L9.21375,2.149875 C9.21375,1.98025455 9.07624545,1.84275 8.906625,1.84275 L7.678125,1.84275 L7.456995,1.39741875 C7.4044565,1.29314865 7.29733818,1.22768745 7.1805825,1.22849252 L7.1805825,1.22849252 Z M2.076165,5.835375 L2.21437125,6.10871625 L2.38329,6.449625 L3.6855,6.449625 L3.6855,7.98525 L1.2285,7.98525 L1.2285,5.835375 L2.076165,5.835375 L2.076165,5.835375 Z M2.2665825,5.22111752 L0.921375,5.22111752 C0.751754546,5.22111752 0.61425,5.35862955 0.61425,5.52825 L0.61425,8.292375 C0.61425,8.46199545 0.751754546,8.5995 0.921375,8.5995 L3.992625,8.5995 C4.16224545,8.5995 4.29975,8.46199545 4.29975,8.292375 L4.29975,6.1425 C4.29975,5.97287955 4.16224545,5.835375 3.992625,5.835375 L2.764125,5.835375 L2.542995,5.39004375 C2.4904565,5.28577365 2.38333818,5.22031245 2.2665825,5.22111752 Z M6.990165,5.835375 L7.12837125,6.10871625 L7.29729,6.449625 L8.5995,6.449625 L8.5995,7.98525 L6.1425,7.98525 L6.1425,5.835375 L6.990165,5.835375 L6.990165,5.835375 Z M7.1805825,5.22111752 L5.835375,5.22111752 C5.66575455,5.22111752 5.52825,5.35862955 5.52825,5.52825 L5.52825,8.292375 C5.52825,8.46199545 5.66575455,8.5995 5.835375,8.5995 L8.906625,8.5995 C9.07624545,8.5995 9.21375,8.46199545 9.21375,8.292375 L9.21375,6.1425 C9.21375,5.97287955 9.07624545,5.835375 8.906625,5.835375 L7.678125,5.835375 L7.456995,5.39004375 C7.4044565,5.28577365 7.29733818,5.22031245 7.1805825,5.22111752 L7.1805825,5.22111752 Z"
-                                id="Fill"
-                            ></path>
-                        </g>
-                    </g>
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_namespace">
-            <g id="Namespace" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.8606 0 0 1.8606 6.8571 6.7143)" stroke="none" fill="#000000">
-                    <path
-                        id="Fill"
-                        d="m2.0762 1.8428 0.13821 0.27334 0.16892 0.34091h1.3022v1.5356h-2.457v-2.1499zm0.19042-0.61426h-1.3452c-0.16962 0-0.30712 0.13751-0.30712 0.30713v2.7641c0 0.16962 0.1375 0.30712 0.30712 0.30712h3.0712c0.16962 0 0.30712-0.1375 0.30712-0.30712v-2.1499c0-0.16962-0.1375-0.30712-0.30712-0.30712h-1.2285l-0.22113-0.44533c-0.052538-0.10427-0.15966-0.16973-0.27641-0.16893zm4.7236 0.61426 0.13821 0.27334 0.16892 0.34091h1.3022v1.5356h-2.457v-2.1499zm0.19042-0.61426h-1.3452c-0.16962 0-0.30712 0.13751-0.30712 0.30713v2.7641c0 0.16962 0.1375 0.30712 0.30712 0.30712h3.0712c0.16962 0 0.30712-0.1375 0.30712-0.30712v-2.1499c0-0.16962-0.1375-0.30712-0.30712-0.30712h-1.2285l-0.22113-0.44533c-0.052538-0.10427-0.15966-0.16973-0.27641-0.16893zm-5.1044 4.6069 0.13821 0.27334 0.16892 0.34091h1.3022v1.5356h-2.457v-2.1499zm0.19042-0.61426h-1.3452c-0.16962 0-0.30712 0.13751-0.30712 0.30713v2.7641c0 0.16962 0.1375 0.30712 0.30712 0.30712h3.0712c0.16962 0 0.30712-0.1375 0.30712-0.30712v-2.1499c0-0.16962-0.1375-0.30712-0.30712-0.30712h-1.2285l-0.22113-0.44533c-0.052538-0.10427-0.15966-0.16973-0.27641-0.16893zm4.7236 0.61426 0.13821 0.27334 0.16892 0.34091h1.3022v1.5356h-2.457v-2.1499zm0.19042-0.61426h-1.3452c-0.16962 0-0.30712 0.13751-0.30712 0.30713v2.7641c0 0.16962 0.1375 0.30712 0.30712 0.30712h3.0712c0.16962 0 0.30712-0.1375 0.30712-0.30712v-2.1499c0-0.16962-0.1375-0.30712-0.30712-0.30712h-1.2285l-0.22113-0.44533c-0.052538-0.10427-0.15966-0.16973-0.27641-0.16893z"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_node">
-            <g id="Node" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.3709 -1.3709 1.3709 1.3709 2.5263 16)" stroke="none" fill="#000000">
-                    <path
-                        id="Fill"
-                        d="m3.3784 3.3784v3.0712h3.0712v-3.0712zm2.457 2.457h-1.8428v-1.8428h1.8428zm3.3784-1.8428v-0.61425h-1.2285v-0.92138c0-0.33924-0.27501-0.61425-0.61425-0.61425h-0.92138v-1.2285h-0.61425v1.2285h-1.8428v-1.2285h-0.61425v1.2285h-0.92138c-0.33924 0-0.61425 0.27501-0.61425 0.61425v0.92138h-1.2285v0.61425h1.2285v1.8428h-1.2285v0.61425h1.2285v0.92138c0 0.33924 0.27501 0.61425 0.61425 0.61425h0.92138v1.2285h0.61425v-1.2285h1.8428v1.2285h0.61425v-1.2285h0.92138c0.33924 0 0.61425-0.27501 0.61425-0.61425v-0.92138h1.2285v-0.61425h-1.2285v-1.8428zm-1.8428 3.3784h-4.914v-4.914h4.914z"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_other">
-            <g id="Other" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.628 0 0 1.628 -9.7143 -9.8571)" stroke="none" fill="#000000">
-                    <g transform="translate(10.881,11.583)">
-                        <path
-                            id="Fill"
-                            d="m9.126 0h-8.424c-0.3877 0-0.702 0.3143-0.702 0.702v7.02c0 0.3877 0.3143 0.702 0.702 0.702h8.424c0.3877 0 0.702-0.3143 0.702-0.702v-7.02c0-0.3877-0.3143-0.702-0.702-0.702zm0 0.702v7.02h-8.424v-7.02z"
-                        />
-                    </g>
-                    <g transform="translate(12.987,15.093)" fillRule="nonzero">
-                        <circle id="Oval" cx="4.914" cy=".61425" r="1" />
-                        <circle cx="2.7641" cy=".61425" r="1" />
-                        <circle cx=".61425" cy=".61425" r="1" />
-                    </g>
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_placement">
-            <g id="Placement" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1.404" />
-                <g transform="matrix(1.8606 0 0 1.8606 6.8571 6.3571)" stroke="none" fill="#000000">
-                    <path
-                        id="Fill"
-                        d="m4.914 7.371-1.8704-2.6413c-0.62492-0.73783-0.76833-1.7699-0.36827-2.6501 0.40006-0.88027 1.2719-1.4509 2.2387-1.4653 1.3669 0.013508 2.4655 1.13 2.457 2.4969 2.512e-4 0.57277-0.19475 1.1285-0.55282 1.5756zm0-6.1425c-1.0284 0.01181-1.853 0.85425-1.8428 1.8827 0.00206 0.4508 0.16424 0.88621 0.45762 1.2285l1.3851 1.9625 1.422-2.0025c0.2703-0.33739 0.41858-0.75627 0.42076-1.1886 0.010234-1.0284-0.81434-1.8709-1.8428-1.8827zm0 2.1499c0.33924 0 0.61425-0.27501 0.61425-0.61425s-0.27501-0.61425-0.61425-0.61425-0.61425 0.27501-0.61425 0.61425 0.27501 0.61425 0.61425 0.61425zm3.6855 0.30712h-0.61425v0.61425h0.61425v4.2998h-7.371v-4.2998h0.61425v-0.61425h-0.61425c-0.33924 0-0.61425 0.27501-0.61425 0.61425v4.2998c0 0.33924 0.27501 0.61425 0.61425 0.61425h7.371c0.33924 0 0.61425-0.27501 0.61425-0.61425v-4.2998c0-0.33924-0.27501-0.61425-0.61425-0.61425z"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_pod">
-            <g id="Pod" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <path
-                    id="Combined-Shape"
-                    stroke="none"
-                    fill="#000000"
-                    d="m22.403 12.142-6.0927-3.3884c-0.19161-0.10656-0.42854-0.10656-0.62015 0l-6.0927 3.3884 6.4027 3.3065zm0.36649 1.1452-6.1538 3.1779v6.8428l5.8485-3.2526c0.18906-0.10514 0.30531-0.29811 0.30531-0.50678zm-7.3846 10.021v-6.8428l-6.1538-3.1779v6.2614c0 0.20867 0.11626 0.40163 0.30531 0.50678zm1.5456-15.568 6.1538 3.4224c0.56717 0.31542 0.91593 0.89432 0.91593 1.5203v6.8658c0 0.626-0.34876 1.2049-0.91593 1.5203l-6.1538 3.4224c-0.57482 0.31968-1.2856 0.31968-1.8604 0l-6.1538-3.4224c-0.56717-0.31543-0.91593-0.89432-0.91593-1.5203v-6.8658c0-0.626 0.34876-1.2049 0.91593-1.5203l6.1538-3.4224c0.57482-0.31968 1.2856-0.31968 1.8604-1e-7z"
-                    fillRule="nonzero"
-                    strokeWidth="1.7532"
-                />
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_policy">
-            <g id="Policy" transform="translate(0)" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g
-                    id="10_Objects/Shield/Shield_outline"
-                    transform="matrix(1.6505 0 0 1.6505 9.3662 7.8192)"
-                    stroke="none"
-                    fill="#000000"
-                >
-                    <path
-                        id="path-policy"
-                        d="m4.054 9.8034c2.0641-3.957e-4 3.7797-1.6066 3.9365-3.6855v-0.04914-4.7707c-9.03e-5 -0.14273-0.097512-0.26658-0.23513-0.29894l-3.6486-0.88042c-0.046511-0.012318-0.095381-0.012318-0.14189 0l-3.6486 0.88042c-0.13535 0.033286-0.23176 0.15417-0.23513 0.29484v4.7707c0.035485 0.58186 0.19891 1.1484 0.47838 1.6585v0.020475c0.70429 1.2805 2.0445 2.0704 3.4946 2.0598zm0-0.61424c-1.079-0.0017774-2.092-0.52553-2.7243-1.4087l6.0162-1.3391c-0.29881 1.5973-1.6826 2.7523-3.2919 2.7478zm-3.3284-7.6495 3.3284-0.819 3.3446 0.819v0.68386l-6.6811 1.4619zm0 2.7887 6.673-1.4619v2.9525l-6.3932 1.4046c-0.17074-0.36954-0.27114-0.76823-0.29595-1.1753z"
-                        fill="#000"
-                        fillRule="evenodd"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_replicaset">
-            <g id="ReplicaSet" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.628 0 0 1.628 6.8573 6.8572)" stroke="none" fill="#000000">
-                    <path
-                        id="Fill"
-                        d="m7.543 4.6034-3.159-1.755c-0.1086-0.0627-0.2424-0.0627-0.351 0l-3.159 1.755c-0.1089 0.0652-0.17453 0.18374-0.17199 0.31064v3.5082c9.2763e-4 0.12737 0.07078 0.24424 0.18252 0.30537l3.159 1.755c0.051282 0.02971 0.10923 0.04601 0.16848 0.04739 0.060333-0.0016 0.11937-0.01782 0.17199-0.04739l3.159-1.755c0.11027-0.0613 0.17875-0.17745 0.17902-0.30362v-3.51c0.0010808-0.12835-0.067424-0.24723-0.17902-0.31064zm-3.682 4.9772-2.457-1.3654v-2.7097l2.457 1.3654zm0.351-3.317-2.4359-1.3496 2.4359-1.3566 2.4359 1.3549zm2.808 1.9516-2.457 1.3654v-2.7097l2.457-1.3654zm3.51-6.1092h-1.404v-1.404h-0.702v1.404h-1.404v0.702h1.404v1.404h0.702v-1.404h1.404z"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_replicationcontroller">
-            <g id="ReplicationController" transform="translate(0)" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.3031 0 0 1.3031 -4.8148 -4.8065)">
-                    <circle
-                        id="path-rc"
-                        cx="15.97"
-                        cy="15.97"
-                        r="5.4405"
-                        stroke="#000"
-                        strokeDasharray="2.106, 2.808"
-                        strokeWidth="1.404"
-                    />
-                    <path
-                        stroke="none"
-                        id="Fill"
-                        d="m18.469 14.339-2.3693-1.3163c-0.08145-0.04702-0.1818-0.04702-0.26325 0l-2.3693 1.3163c-0.08167 0.0489-0.1309 0.1378-0.12899 0.23298v2.6312c6.96e-4 0.09552 0.05309 0.18318 0.13689 0.22903l2.3693 1.3163c0.03846 0.02228 0.08192 0.03451 0.12636 0.03554 0.04525-0.0012 0.08953-0.01337 0.12899-0.03554l2.3693-1.3163c0.0827-0.04598 0.13406-0.13309 0.13427-0.22771v-2.6325c8.11e-4 -0.09626-0.05057-0.18542-0.13427-0.23298zm-2.7615 3.7329-1.8428-1.024v-2.0323l1.8428 1.024zm0.26325-2.4877-1.827-1.0122 1.827-1.0175 1.827 1.0162zm2.106 1.4637-1.8428 1.024v-2.0323l1.8428-1.024z"
-                        fill="#000"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_route">
-            <g transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.6422 0 0 1.6422 8 8.5)" stroke="none" fill="#000000">
-                    <g transform="translate(0 .702)">
-                        <path
-                            id="Fill"
-                            d="m6.2542 7.5465h-4.0205c-1.1102 0-2.0103-0.90003-2.0103-2.0103 0-1.1102 0.90003-2.0103 2.0103-2.0103h0.67009v0.67009h-0.67009c-0.74016 0-1.3402 0.60002-1.3402 1.3402 0 0.74016 0.60002 1.3402 1.3402 1.3402h4.0205zm1.3402-3.3505h-0.67009v-0.67009h0.67009c0.74016 0 1.3402-0.60002 1.3402-1.3402s-0.60002-1.3402-1.3402-1.3402h-4.0205v-0.67009h4.0205c1.1102 0 2.0103 0.90003 2.0103 2.0103 0 1.1102-0.90003 2.0103-2.0103 2.0103z"
-                        />
-                    </g>
-                    <path
-                        id="Combined-Shape"
-                        d="m1.3615 0c0.75191 0 1.3615 0.60954 1.3615 1.3615 0 0.36108-0.14344 0.70737-0.39876 0.96269-0.25532 0.25532-0.60161 0.39876-0.96269 0.39876-0.75191 0-1.3615-0.60954-1.3615-1.3615 0-0.75191 0.60954-1.3615 1.3615-1.3615zm0 0.68073c-0.37596 0-0.68073 0.30477-0.68073 0.68073 0 0.37596 0.30477 0.68073 0.68073 0.68073 0.37596 0 0.68073-0.30477 0.68073-0.68073 0-0.37596-0.30477-0.68073-0.68073-0.68073z"
-                    />
-                    <path
-                        id="Combined-Shape-Copy"
-                        d="m4.8715 3.159c0.75191 0 1.3615 0.60954 1.3615 1.3615 0 0.36108-0.14344 0.70737-0.39876 0.96269-0.25532 0.25532-0.60161 0.39876-0.96269 0.39876-0.75191 0-1.3615-0.60954-1.3615-1.3615 0-0.75191 0.60954-1.3615 1.3615-1.3615zm0 0.68073c-0.37596 0-0.68073 0.30477-0.68073 0.68073s0.30477 0.68073 0.68073 0.68073 0.68073-0.30477 0.68073-0.68073-0.30477-0.68073-0.68073-0.68073z"
-                    />
-                    <path d="m8.3815 6.318c0.75191 0 1.3615 0.60954 1.3615 1.3615 0 0.36108-0.14344 0.70737-0.39876 0.96269-0.25532 0.25532-0.60161 0.39876-0.96269 0.39876-0.75191 0-1.3615-0.60954-1.3615-1.3615 0-0.75191 0.60954-1.3615 1.3615-1.3615zm0 0.68073c-0.37596 0-0.68073 0.30477-0.68073 0.68073s0.30477 0.68073 0.68073 0.68073c0.37596 0 0.68073-0.30477 0.68073-0.68073s-0.30477-0.68073-0.68073-0.68073z" />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_placements">
-            <g transform="translate(0)" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.8993 0 0 1.8993 8.6667 8.6667)">
-                    <g stroke="none" fill="#000000" fillRule="nonzero">
-                        <g id="icon/file/template/32-copy" transform="translate(.96525 .96525)">
-                            <polygon
-                                id="Polygon"
-                                points="0.086213 0.32175 0.6435 0 1.2008 0.32175 1.2008 0.96525 0.6435 1.287 0.086213 0.96525"
-                            />
-                        </g>
-                        <g transform="translate(5.4698 .96525)">
-                            <polygon points="0.086213 0.32175 0.6435 0 1.2008 0.32175 1.2008 0.96525 0.6435 1.287 0.086213 0.96525" />
-                        </g>
-                        <g transform="translate(3.2175 .96525)">
-                            <polygon points="0.086213 0.32175 0.6435 0 1.2008 0.32175 1.2008 0.96525 0.6435 1.287 0.086213 0.96525" />
-                        </g>
-                        <g transform="translate(.96525 3.2175)">
-                            <polygon points="0.086213 0.32175 0.6435 0 1.2008 0.32175 1.2008 0.96525 0.6435 1.287 0.086213 0.96525" />
-                        </g>
-                        <g transform="translate(5.4698 3.2175)">
-                            <polygon points="0.086213 0.32175 0.6435 0 1.2008 0.32175 1.2008 0.96525 0.6435 1.287 0.086213 0.96525" />
-                        </g>
-                        <g transform="translate(3.2175,3.2175)">
-                            <polygon points="0.086213 0.32175 0.6435 0 1.2008 0.32175 1.2008 0.96525 0.6435 1.287 0.086213 0.96525" />
-                        </g>
-                        <g transform="translate(.96525 5.4698)">
-                            <polygon points="0.086213 0.32175 0.6435 0 1.2008 0.32175 1.2008 0.96525 0.6435 1.287 0.086213 0.96525" />
-                        </g>
-                        <g transform="translate(5.4698 5.4698)">
-                            <polygon points="0.086213 0.32175 0.6435 0 1.2008 0.32175 1.2008 0.96525 0.6435 1.287 0.086213 0.96525" />
-                        </g>
-                        <g transform="translate(3.2175 5.4698)">
-                            <polygon points="0.086213 0.32175 0.6435 0 1.2008 0.32175 1.2008 0.96525 0.6435 1.287 0.086213 0.96525" />
-                        </g>
-                    </g>
-                    <rect id="Rectangle-Copy" width="7.722" height="7.722" rx=".702" stroke="#000" strokeWidth=".702" />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_secret">
-            <g id="Secret" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.9509 0 0 1.9509 6.8587 6.8589)" stroke="none" fill="#000000">
-                    <path
-                        id="Fill"
-                        d="m6.1425 0.58496c-1.4539 0-2.6325 1.1786-2.6325 2.6325-2.548e-4 0.25874 0.038196 0.51606 0.11408 0.76342l-3.0391 3.0391v1.755h1.755l3.0391-3.0391c0.76221 0.23387 1.5893 0.1097 2.2493-0.33765 0.65995-0.44736 1.0816-1.1697 1.1466-1.9643 0.059053-0.62433-0.11264-1.2487-0.48262-1.755-0.49757-0.69084-1.2985-1.0984-2.1499-1.094zm0 4.68c-0.19822-7.785e-4 -0.39528-0.030337-0.585-0.08775l-0.33638-0.10238-1.1788 1.1788-0.41242-0.40365-0.41242 0.41242 0.40365 0.40365-0.46508 0.46508-0.40365-0.40365-0.41242 0.41242 0.40365 0.40365-0.64642 0.64642h-0.92722v-0.92722l2.8665-2.8753 0.24862-0.24862-0.084825-0.27788c-0.30367-0.95492 0.11368-1.9909 0.9945-2.4687 0.62531-0.32689 1.3749-0.30878 1.9837 0.047938s0.99102 1.0018 1.0115 1.7071c0.020298 0.41392-0.091339 0.8236-0.31882 1.17-0.37621 0.59077-1.0283 0.94826-1.7287 0.9477zm0.2925-1.755c0.32309 0 0.585-0.26191 0.585-0.585s-0.26191-0.585-0.585-0.585-0.585 0.26191-0.585 0.585 0.26191 0.585 0.585 0.585z"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_service">
-            <g id="Service" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.924 0 0 1.924 8.5 8.5)" fillRule="nonzero" stroke="none" fill="#000000">
-                    <circle cx="1.8715" cy="1.4524" r="1" />
-                    <path
-                        id="Combined-Shape"
-                        d="m1.2924 5.8097v0.16771l0.71339 0.41326v1.1619l-1.0029 0.58097-1.0029-0.58097v-1.1619l0.71339-0.41326v-1.3296c0-0.32086 0.25924-0.58097 0.57903-0.58097h2.3161v-1.1619h-2.6056c-0.31979 0-0.57903-0.26011-0.57903-0.58097v-1.7429c0-0.32086 0.25924-0.58097 0.57903-0.58097h5.7903c0.31979 0 0.57903 0.26011 0.57903 0.58097v1.7429c0 0.32086-0.25924 0.58097-0.57903 0.58097h-2.6056v1.1619h2.3161c0.31979 0 0.57903 0.26011 0.57903 0.58097v1.3296l0.71339 0.41326v1.1619l-1.0029 0.58097-1.0029-0.58097v-1.1619l0.71339-0.41326v-1.3296h-2.3161v1.3296l0.71339 0.41326v1.1619l-1.0029 0.58097-1.0029-0.58097v-1.1619l0.71339-0.41326v-1.3296h-2.3161zm-0.38328 2.3239h0.18752c-0.030953 0.0025-0.062224 0.00377-0.093761 0.00377s-0.062808-0.00127-0.093761-0.00377zm1.8308 0h2.3161zm3.9594 0h0.18752c-0.030953 0.0025-0.062224 0.00377-0.093761 0.00377s-0.062808-0.00127-0.093761-0.00377zm-5.6965-5.8097h5.7903v-1.7429h-5.7903zm0 4.1566-0.42388 0.24554v0.49109l0.42388 0.24554 0.42388-0.24554v-0.49109zm2.4713 0.24554v0.49109l0.42388 0.24554 0.42388-0.24554v-0.49109l-0.42388-0.24554zm2.8951 0v0.49109l0.42388 0.24554 0.42388-0.24554v-0.49109l-0.42388-0.24554z"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_statefulset">
-            <g id="StatefulSet" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g
-                    id="Statefulset"
-                    transform="matrix(1.6437 0 0 1.6437 9.8454 8)"
-                    fillRule="nonzero"
-                    stroke="none"
-                    fill="#000000"
-                >
-                    <g id="Group" transform="translate(1.755)">
-                        <polygon id="Path" points="0 6.318 5.616 6.318 5.616 7.02 0 7.02" />
-                        <polygon points="0 4.563 2.457 4.563 2.457 5.265 0 5.265" />
-                        <polygon id="Path-Copy" points="0 8.073 5.616 8.073 5.616 8.775 0 8.775" />
-                        <polygon points="3.159 1.3443 3.159 5.616 3.861 5.616 3.861 1.3443 5.1211 2.6009 5.616 2.106 3.51 0 1.404 2.106 1.8989 2.6009" />
-                    </g>
-                    <polygon
-                        id="Polygon-Copy"
-                        points="0.702 5.967 1.3099 6.318 1.3099 7.02 0.702 7.371 0.09405 7.02 0.09405 6.318"
-                    />
-                    <polygon points="0.702 7.722 1.3099 8.073 1.3099 8.775 0.702 9.126 0.09405 8.775 0.09405 8.073" />
-                    <polygon points="0.702 4.212 1.3099 4.563 1.3099 5.265 0.702 5.616 0.09405 5.265 0.09405 4.563" />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_storageclass">
-            <g id="StorageClass" transform="translate(0)" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g
-                    id="01_Hardware/Storage-stack/Storage-stack_outline"
-                    transform="matrix(1.7145 0 0 1.7145 8.3077 7.3085)"
-                    stroke="none"
-                    fill="#000000"
-                >
-                    <path
-                        id="path-sc"
-                        d="m4.4867 0.11168c-2.4542 0-4.3745 1.1302-4.3745 2.5687v4.4673c0 1.4385 1.9203 2.5687 4.3745 2.5687 2.4542 0 4.3745-1.1302 4.3745-2.5687v-4.4673c0-1.4385-1.9203-2.5687-4.3745-2.5687zm3.7015 4.8023c0 1.0275-1.696 1.8986-3.7015 1.8986-2.0056 0-3.7015-0.87112-3.7015-1.8986v-0.84878c0.76722 0.71476 2.1222 1.1838 3.7015 1.1838 1.5793 0 2.9343-0.46906 3.7015-1.1838zm-3.7015-4.1322c2.0056 0 3.7015 0.87112 3.7015 1.8986 0 1.0275-1.696 1.8986-3.7015 1.8986-2.0056 0-3.7015-0.87112-3.7015-1.8986 0-1.0275 1.696-1.8986 3.7015-1.8986zm0 8.2645c-2.0056 0-3.7015-0.87112-3.7015-1.8986v-0.84878c0.76722 0.71476 2.1222 1.1838 3.7015 1.1838 1.5793 0 2.9343-0.46906 3.7015-1.1838v0.84878c0 1.0275-1.696 1.8986-3.7015 1.8986z"
-                        fill="#000"
-                        fillRule="evenodd"
-                    />
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_subscription">
-            <g transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.7034 0 0 1.7034 7.3454 7.5036)">
-                    <path
-                        id="Fill"
-                        d="m4.6069 1.8428h3.9926v0.61425h-3.9926zm0 5.5282h3.9926v0.61425h-3.9926zm-3.3784-2.7641h3.9926v0.61425h-3.9926zm0.92138-1.2285c-0.67848 0-1.2285-0.55002-1.2285-1.2285s0.55002-1.2285 1.2285-1.2285 1.2285 0.55002 1.2285 1.2285c0 0.32582-0.12943 0.63829-0.35982 0.86868-0.23039 0.23039-0.54286 0.35982-0.86868 0.35982zm0-1.8428c-0.33924 0-0.61425 0.27501-0.61425 0.61425s0.27501 0.61425 0.61425 0.61425 0.61425-0.27501 0.61425-0.61425-0.27501-0.61425-0.61425-0.61425zm0 7.371c-0.67848 0-1.2285-0.55002-1.2285-1.2285s0.55002-1.2285 1.2285-1.2285 1.2285 0.55002 1.2285 1.2285-0.55002 1.2285-1.2285 1.2285zm0-1.8428c-0.33924 0-0.61425 0.27501-0.61425 0.61425s0.27501 0.61425 0.61425 0.61425 0.61425-0.27501 0.61425-0.61425-0.27501-0.61425-0.61425-0.61425zm5.5282-0.92138c-0.67848 0-1.2285-0.55002-1.2285-1.2285s0.55002-1.2285 1.2285-1.2285 1.2285 0.55002 1.2285 1.2285c0 0.32582-0.12943 0.63829-0.35982 0.86868-0.23039 0.23039-0.54286 0.35982-0.86868 0.35982zm0-1.8428c-0.33924 0-0.61425 0.27501-0.61425 0.61425s0.27501 0.61425 0.61425 0.61425 0.61425-0.27501 0.61425-0.61425-0.27501-0.61425-0.61425-0.61425z"
-                    />
-                </g>
-                <path
-                    d="m14.794 30.14c-6.2113-0.48184-11.307-4.7125-12.909-10.718-0.13011-0.48758-0.2822-1.2824-0.368-1.9231-0.08417-0.62852-0.097424-2.6239-0.021513-3.2388 0.36714-2.9741 1.4643-5.5314 3.3261-7.7528 0.46385-0.55344 1.3556-1.4226 1.9432-1.894 2.1664-1.7378 4.6752-2.781 7.4957-3.1166 0.57422-0.068343 2.5143-0.069174 3.0655-0.00131 3.5858 0.44144 6.6361 1.9759 9.0133 4.534 2.1307 2.293 3.4114 5.1381 3.767 8.3691 0.06456 0.58657 0.06494 2.1782 6.65e-4 2.7886-0.29746 2.8253-1.3557 5.4217-3.1 7.6063-0.77479 0.97035-2.0069 2.1172-3.0357 2.8256-2.0089 1.3832-4.3413 2.2395-6.7639 2.483-0.59038 0.05936-1.8776 0.07976-2.4129 0.03824zm-3.4413-7.4771c0.14556-0.02334 0.39698-0.10869 0.57949-0.1967 0.73433-0.35412 1.1743-1.0661 1.1725-1.8974-0.0013-0.58821-0.19318-1.051-0.60931-1.4693-0.81602-0.82028-2.0796-0.83502-2.9357-0.03424-0.98221 0.91881-0.83971 2.5229 0.2912 3.2781 0.44591 0.29777 0.95811 0.40676 1.5018 0.31958zm10.64-2.0744v-0.53399h-6.8035v1.068h6.8035zm-0.98888-2.6949c0.77869-0.2191 1.3806-0.90343 1.5054-1.7114 0.13428-0.86976-0.31201-1.7512-1.0991-2.1707-0.81511-0.43445-1.8035-0.2812-2.4673 0.38253-0.64986 0.64986-0.8077 1.5718-0.41041 2.3972 0.44762 0.92994 1.4675 1.3849 2.4714 1.1024zm-4.7466-2.0122v-0.53399h-6.843v1.068h6.843zm-4.6675-2.6959c0.33478-0.09265 0.64581-0.27655 0.90344-0.53418 0.84842-0.84842 0.81539-2.2531-0.07174-3.0507-0.83973-0.75499-2.1392-0.70828-2.9201 0.10499-1.0471 1.0905-0.65934 2.8738 0.74527 3.4274 0.40926 0.16132 0.88306 0.17984 1.3431 0.05251zm10.403-2.0112v-0.53399h-6.8035v1.068h6.8035z"
-                    fill="#fff"
-                    stroke="none"
-                />
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 32 32" id="diagramShapes_subscriptionblocked">
-            <g id="SubscriptionBlocked" transform="translate(0)" fill="none" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1" />
-                <g transform="matrix(1.4984 0 0 1.4984 -8.013 -6.858)" stroke="none">
-                    <g transform="translate(11.713 11.713)" fill="#000">
-                        <path
-                            id="Fill"
-                            d="m3.6547 1.1427h4.3146v0.6206h-4.3146zm0.28071 5.5854h4.0339v0.6206h-4.0339zm-3.4133-2.7927h4.3146v0.6206h-4.3146zm0.9309-1.2412c-0.6855 0-1.2412-0.5557-1.2412-1.2412 0-0.6855 0.5557-1.2412 1.2412-1.2412 0.6855 0 1.2412 0.5557 1.2412 1.2412 0 0.32919-0.13077 0.64489-0.36354 0.87766s-0.54847 0.36354-0.87766 0.36354zm0-1.8618c-0.34275 0-0.6206 0.27785-0.6206 0.6206s0.27785 0.6206 0.6206 0.6206 0.6206-0.27785 0.6206-0.6206-0.27785-0.6206-0.6206-0.6206zm0 7.4472c-0.6855 0-1.2412-0.5557-1.2412-1.2412 0-0.6855 0.5557-1.2412 1.2412-1.2412 0.6855 0 1.2412 0.5557 1.2412 1.2412 0 0.6855-0.5557 1.2412-1.2412 1.2412zm0-1.8618c-0.34275 0-0.6206 0.27785-0.6206 0.6206 0 0.34275 0.27785 0.6206 0.6206 0.6206s0.6206-0.27785 0.6206-0.6206c0-0.34275-0.27785-0.6206-0.6206-0.6206zm5.5854-0.9309c-0.6855 0-1.2412-0.5557-1.2412-1.2412s0.5557-1.2412 1.2412-1.2412c0.6855 0 1.2412 0.5557 1.2412 1.2412 0 0.32919-0.13077 0.64489-0.36354 0.87766-0.23277 0.23277-0.54847 0.36354-0.87766 0.36354zm0-1.8618c-0.34275 0-0.6206 0.27785-0.6206 0.6206 0 0.34275 0.27785 0.6206 0.6206 0.6206 0.34275 0 0.6206-0.27785 0.6206-0.6206 0-0.34275-0.27785-0.6206-0.6206-0.6206z"
-                        />
-                    </g>
-                    <g transform="rotate(-45 18.612 -7.8762)">
-                        <rect id="Rectangle" width="1" height="13.303" fill="#000" />
-                        <rect id="Rectangle-Copy" x=".7191" width="1.0787" height="13.303" fill="#fff" />
-                    </g>
-                </g>
-            </g>
-        </symbol>
-        <symbol viewBox="0 0 406 149" id="diagramShapes_legend">
-            <defs>
-                <circle id="path-1" cx="47.6343391" cy="46.9487385" r="46.9487385"></circle>
-                <filter
-                    x="-6.9%"
-                    y="-5.9%"
-                    width="113.8%"
-                    height="113.8%"
-                    filterUnits="objectBoundingBox"
-                    id="filter-2"
-                >
-                    <feOffset dx="0" dy="1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-                    <feGaussianBlur stdDeviation="2" in="shadowOffsetOuter1" result="shadowBlurOuter1"></feGaussianBlur>
-                    <feColorMatrix
-                        values="0 0 0 0 0   0 0 0 0 0   0 0 0 0 0  0 0 0 0.2 0"
-                        type="matrix"
-                        in="shadowBlurOuter1"
-                    ></feColorMatrix>
-                </filter>
-                <rect id="path-3" x="0" y="100.157309" width="94.2253729" height="39.8020972" rx="3.1299159"></rect>
-                <filter
-                    x="-8.5%"
-                    y="-17.6%"
-                    width="117.0%"
-                    height="140.2%"
-                    filterUnits="objectBoundingBox"
-                    id="filter-4"
-                >
-                    <feOffset dx="0" dy="1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-                    <feGaussianBlur
-                        stdDeviation="2.5"
-                        in="shadowOffsetOuter1"
-                        result="shadowBlurOuter1"
-                    ></feGaussianBlur>
-                    <feComposite
-                        in="shadowBlurOuter1"
-                        in2="SourceAlpha"
-                        operator="out"
-                        result="shadowBlurOuter1"
-                    ></feComposite>
-                    <feColorMatrix
-                        values="0 0 0 0 0   0 0 0 0 0   0 0 0 0 0  0 0 0 0.2 0"
-                        type="matrix"
-                        in="shadowBlurOuter1"
-                    ></feColorMatrix>
-                </filter>
-                <circle id="path-5" cx="12.5196636" cy="12.5196636" r="12.5196636"></circle>
-                <filter
-                    x="-31.9%"
-                    y="-28.0%"
-                    width="163.9%"
-                    height="163.9%"
-                    filterUnits="objectBoundingBox"
-                    id="filter-6"
-                >
-                    <feOffset dx="0" dy="1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-                    <feGaussianBlur
-                        stdDeviation="2.5"
-                        in="shadowOffsetOuter1"
-                        result="shadowBlurOuter1"
-                    ></feGaussianBlur>
-                    <feColorMatrix
-                        values="0 0 0 0 0   0 0 0 0 0   0 0 0 0 0  0 0 0 0.3 0"
-                        type="matrix"
-                        in="shadowBlurOuter1"
-                    ></feColorMatrix>
-                </filter>
-            </defs>
-            <g id="Topology" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-                <g id="App-Topology--How-to-read-this-01.00" transform="translate(-924.000000, -1613.000000)">
-                    <g id="Group-5" transform="translate(924.000000, 1616.000000)">
-                        <g id="Group-13-Copy-2" transform="translate(149.000000, 0.000000)">
-                            <g id="Oval-Copy-8">
-                                <use fill="black" fillOpacity="1" filter="url(#filter-2)" href="#path-1"></use>
-                                <use fill="#FFFFFF" fillRule="evenodd" href="#path-1"></use>
-                            </g>
-                            <circle
-                                id="Oval-Copy-9"
-                                stroke="#c7c7c7"
-                                strokeWidth="4.1732212"
-                                fill="#FFFFFF"
-                                cx="47.6343391"
-                                cy="46.9487385"
-                                r="40.6889067"
-                            ></circle>
-                            <g
-                                id="Organization-/-Data-/-data-vis--1-/-24"
-                                transform="translate(33.028065, 33.385770)"
-                                fill="#000000"
-                            >
-                                <path
-                                    d="M23.7351956,20.083627 C23.0851254,20.0870047 22.4480264,20.265834 21.8911534,20.6012369 L18.4102957,17.1194662 C19.4243925,15.597454 19.4243925,13.6150944 18.4102957,12.0930821 L20.735432,9.76794587 C21.9499059,10.3529292 23.4092284,9.97658951 24.1893331,8.87723099 C24.9694378,7.77787248 24.842799,6.27613474 23.8896063,5.32294208 C22.9364137,4.36974942 21.4346759,4.24311057 20.3353174,5.02321528 C19.2359589,5.80331998 18.8596191,7.26264245 19.4446025,8.47711639 L17.1194662,10.8022527 C15.597454,9.78815586 13.6150944,9.78815586 12.0930821,10.8022527 L8.61131153,7.32048205 C8.94655622,6.76387497 9.12537951,6.1271127 9.12892137,5.47735282 C9.12892137,3.4606472 7.49405845,1.82578427 5.47735282,1.82578427 C3.4606472,1.82578427 1.82578427,3.4606472 1.82578427,5.47735282 C1.82578427,7.49405845 3.4606472,9.12892137 5.47735282,9.12892137 C6.12742296,9.12554365 6.76452195,8.94671438 7.32139494,8.61131153 L10.8022527,12.0930821 C9.78815586,13.6150944 9.78815586,15.597454 10.8022527,17.1194662 L7.32048205,20.6012369 C6.76385689,20.266034 6.12710636,20.087214 5.47735282,20.083627 C3.4606472,20.083627 1.82578427,21.7184899 1.82578427,23.7351956 C1.82578427,25.7519012 3.4606472,27.3867641 5.47735282,27.3867641 C7.49405845,27.3867641 9.12892137,25.7519012 9.12892137,23.7351956 C9.12554365,23.0851254 8.94671438,22.4480264 8.61131153,21.8911534 L12.0930821,18.4102957 C12.5785764,18.7338856 13.1219377,18.960777 13.6933821,19.0785328 L13.6933821,22.0773834 C12.4358305,22.5219951 11.6814846,23.8079222 11.9070387,25.1225479 C12.1325928,26.4371736 13.2724394,27.3981087 14.6062742,27.3981087 C15.940109,27.3981087 17.0799556,26.4371736 17.3055097,25.1225479 C17.5310638,23.8079222 16.7767179,22.5219951 15.5191663,22.0773834 L15.5191663,19.0785328 C16.0906107,18.960777 16.633972,18.7338856 17.1194662,18.4102957 L20.6012369,21.8920663 C20.2659922,22.4486734 20.0871689,23.0854357 20.083627,23.7351956 C20.083627,25.7519012 21.7184899,27.3867641 23.7351956,27.3867641 C25.7519012,27.3867641 27.3867641,25.7519012 27.3867641,23.7351956 C27.3867641,21.7184899 25.7519012,20.083627 23.7351956,20.083627 L23.7351956,20.083627 Z M14.6062742,11.8675978 C16.1188034,11.8675978 17.3449506,13.093745 17.3449506,14.6062742 C17.3449506,16.1188034 16.1188034,17.3449506 14.6062742,17.3449506 C13.093745,17.3449506 11.8675978,16.1188034 11.8675978,14.6062742 C11.8691074,13.0943707 13.0943707,11.8691074 14.6062742,11.8675978 L14.6062742,11.8675978 Z M3.65156855,5.47735282 C3.65156855,4.46900001 4.46900001,3.65156855 5.47735282,3.65156855 C6.48570563,3.65156855 7.3031371,4.46900001 7.3031371,5.47735282 C7.3031371,6.48570563 6.48570563,7.3031371 5.47735282,7.3031371 C4.46941718,7.30213067 3.65257498,6.48528846 3.65156855,5.47735282 L3.65156855,5.47735282 Z M5.47735282,25.5609798 C4.46900001,25.5609798 3.65156855,24.7435484 3.65156855,23.7351956 C3.65156855,22.7268428 4.46900001,21.9094113 5.47735282,21.9094113 C6.48570563,21.9094113 7.3031371,22.7268428 7.3031371,23.7351956 C7.30213067,24.7431312 6.48528846,25.5599734 5.47735282,25.5609798 L5.47735282,25.5609798 Z M23.7351956,25.5609798 C22.7268428,25.5609798 21.9094113,24.7435484 21.9094113,23.7351956 C21.9094113,22.7268428 22.7268428,21.9094113 23.7351956,21.9094113 C24.7435484,21.9094113 25.5609798,22.7268428 25.5609798,23.7351956 C25.5594716,24.742923 24.742923,25.5594716 23.7351956,25.5609798 L23.7351956,25.5609798 Z"
-                                    id="Fill"
-                                ></path>
-                            </g>
-                            <g id="Rectangle-Copy-2" fillRule="nonzero">
-                                <use fill="black" fillOpacity="1" filter="url(#filter-4)" href="#path-3"></use>
-                                <use fillOpacity="0.9" fill="#FFFFFF" href="#path-3"></use>
-                            </g>
-                            <text
-                                id="Cluster-Possiblerept"
-                                fontFamily="RedHatText-Medium, Red Hat Text"
-                                fontSize="11.3720278"
-                                fontWeight="400"
-                                fill="#152935"
-                            >
-                                <tspan x="9.97101779" y="117.142818">
-                                    Cluster
-                                </tspan>
-                                <tspan
-                                    x="48.8178646"
-                                    y="117.142818"
-                                    fontFamily="RedHatText-Regular, Red Hat Text"
-                                    fontWeight="normal"
-                                ></tspan>
-                                <tspan
-                                    x="9.97101779"
-                                    y="132.142818"
-                                    fontFamily="RedHatText-Regular, Red Hat Text"
-                                    fontWeight="normal"
-                                >
-                                    Possiblereptile
-                                </tspan>
-                            </text>
-                            <g id="Group-8-Copy" transform="translate(0.685601, 66.771539)">
-                                <g id="Oval">
-                                    <use fill="black" fillOpacity="1" filter="url(#filter-6)" href="#path-5"></use>
-                                    <use fill="#FFFFFF" fillRule="evenodd" href="#path-5"></use>
-                                </g>
-                                <g id="Icons-/-14px-/-check-circle" transform="translate(5.216526, 5.216526)">
-                                    <path
-                                        d="M14.6062742,7.3031371 C14.6062742,8.61848721 14.2774366,9.83567657 13.6197617,10.9547056 C12.9620869,12.0737347 12.0737347,12.9620869 10.9547056,13.6197617 C9.83567657,14.2774366 8.61848721,14.6062742 7.3031371,14.6062742 C5.98778699,14.6062742 4.77059762,14.2774366 3.65156855,13.6197617 C2.53253948,12.9620869 1.6441873,12.0737347 0.986512471,10.9547056 C0.32883764,9.83567657 0,8.61848721 0,7.3031371 C0,5.98778699 0.32883764,4.77059762 0.986512471,3.65156855 C1.6441873,2.53253948 2.53253948,1.6441873 3.65156855,0.986512471 C4.77059762,0.32883764 5.98778699,0 7.3031371,0 C8.61848721,0 9.83567657,0.32883764 10.9547056,0.986512471 C12.0737347,1.6441873 12.9620869,2.53253948 13.6197617,3.65156855 C14.2774366,4.77059762 14.6062742,5.98778699 14.6062742,7.3031371 Z"
-                                        id="check-circle"
-                                        stroke="#108926"
-                                        strokeWidth="1.0433053"
-                                    ></path>
-                                    <path
-                                        d="M6.44914123,11.1608426 L11.8675978,5.74238602 C11.9657581,5.66385752 12.0148385,5.55588133 12.0148385,5.41845656 C12.0148385,5.28103178 11.9657581,5.17305559 11.8675978,5.09452709 L11.2197388,4.41722002 C11.1215786,4.31905972 11.0086937,4.26997935 10.8810853,4.26997935 C10.7534769,4.26997935 10.6405921,4.31905972 10.5424318,4.41722002 L6.12521176,8.83444004 L4.06384242,6.77307069 C3.96568212,6.6749104 3.85279731,6.62583003 3.72518888,6.62583003 C3.59758045,6.62583003 3.48469564,6.6749104 3.38653535,6.77307069 L2.73867641,7.45037776 C2.64051612,7.52890627 2.59143574,7.63688246 2.59143574,7.77430723 C2.59143574,7.911732 2.64051612,8.01970819 2.73867641,8.0982367 L5.80128229,11.1608426 C5.8798108,11.2590029 5.98778699,11.3080832 6.12521176,11.3080832 C6.26263653,11.3080832 6.37061272,11.2590029 6.44914123,11.1608426 Z"
-                                        id="Path"
-                                        fill="#3E8635"
-                                    ></path>
-                                </g>
-                            </g>
-                        </g>
-                        <g id="Group-Copy-7" transform="translate(93.000000, 126.000000)">
-                            <line
-                                x1="0.5"
-                                y1="2.5"
-                                x2="61.7500645"
-                                y2="2.02790668"
-                                id="Line-Copy"
-                                stroke="#979797"
-                                strokeLinecap="square"
-                            ></line>
-                            <circle id="Oval-Copy-15" fill="#979797" cx="61.25" cy="2" r="2"></circle>
-                        </g>
-                        <g id="Group-Copy-8" transform="translate(86.000000, 48.000000)">
-                            <line
-                                x1="0.5"
-                                y1="2.5"
-                                x2="88.7492627"
-                                y2="1.81980663"
-                                id="Line-Copy"
-                                stroke="#979797"
-                                strokeLinecap="square"
-                            ></line>
-                            <circle id="Oval-Copy-15" fill="#979797" cx="90.25" cy="2" r="2"></circle>
-                        </g>
-                        <g id="Group-Copy-9" transform="translate(126.000000, 78.000000)">
-                            <line
-                                x1="-1.34295963e-12"
-                                y1="2.5"
-                                x2="18.2513419"
-                                y2="2.35932526"
-                                id="Line-Copy"
-                                stroke="#979797"
-                                strokeLinecap="square"
-                            ></line>
-                            <circle id="Oval-Copy-15" fill="#979797" cx="18" cy="2" r="2"></circle>
-                        </g>
-                        <line
-                            x1="205.75"
-                            y1="113.25"
-                            x2="304.125007"
-                            y2="113.363466"
-                            id="Line-Copy-9"
-                            stroke="#979797"
-                            strokeLinecap="square"
-                        ></line>
-                        <text
-                            id="Resource-icon-Copy-2"
-                            fontFamily="RedHatText-Medium, Red Hat Text"
-                            fontSize="12"
-                            fontWeight="400"
-                            fill="#5A6872"
-                        >
-                            <tspan x="0" y="54">
-                                Resource icon
-                            </tspan>
-                        </text>
-                        <text
-                            id="Cluster-deploy-statu-Copy-2"
-                            fontFamily="RedHatText-Medium, Red Hat Text"
-                            fontSize="12"
-                            fontWeight="400"
-                            fill="#5A6872"
-                        >
-                            <tspan x="0" y="83">
-                                Resource deploy status
-                            </tspan>
-                        </text>
-                        <text
-                            id="Resource-type-Copy-2"
-                            fontFamily="RedHatText-Medium, Red Hat Text"
-                            fontSize="12"
-                            fontWeight="400"
-                            fill="#5A6872"
-                        >
-                            <tspan x="313" y="118">
-                                Resource type
-                            </tspan>
-                        </text>
-                        <text
-                            id="Resource-name-Copy-2"
-                            fontFamily="RedHatText-Medium, Red Hat Text"
-                            fontSize="12"
-                            fontWeight="400"
-                            fill="#5A6872"
-                        >
-                            <tspan x="0" y="131">
-                                Resource name
-                            </tspan>
-                        </text>
-                        <circle id="Oval-Copy-19" fill="#979797" cx="205" cy="113" r="2"></circle>
-                    </g>
-                </g>
-            </g>
-        </symbol>
-    </svg>
-)
