+++ frontend/src/routes/Applications/ApplicationDetails/ApplicationOverview/ApplicationOverview.tsx
@@ -61,7 +61,7 @@ import { getDiagramElements } from '../ApplicationTopology/model/topology'
 import { getAuthorizedNamespaces, rbacCreate } from '../../../../lib/rbac-util'
 import { Link } from 'react-router-dom'
 import { useAllClusters } from '../../../Infrastructure/Clusters/ManagedClusters/components/useAllClusters'
-import { DiagramIcons } from '../../../../components/Topology/shapes/DiagramIcons'
+import { DrawerShapes } from '../ApplicationTopology/components/DrawerShapes'
 import { useRecoilState } from '../../../../shared-recoil'
 import { PluginContext } from '../../../../lib/PluginContext'
 
@@ -100,7 +100,7 @@ export function ApplicationOverviewPageContent(props: { applicationData: Applica
         open: false,
     })
     const [hasSyncPermission, setHasSyncPermission] = useState(false)
-    const openTabIcon = '#diagramIcons_open-new-tab'
+    const openTabIcon = '#drawerShapes_open-new-tab'
 
     let isArgoApp = false
     let isAppSet = false
@@ -351,7 +351,7 @@ export function ApplicationOverviewPageContent(props: { applicationData: Applica
     return (
         <AcmPageContent id="overview">
             <SyncResourceModal {...modalProps} />
-            <DiagramIcons />
+            <DrawerShapes />
             <PageSection>
                 <div className="overview-cards-container">
                     <AcmDescriptionList
