+++ frontend/src/routes/Applications/ApplicationDetails/ApplicationTopology/options/utilities.js
